chore: golangci-lint clean + Internet as documented default scheme
- Default LoadBalancer scheme documented as 'Internet' (matches the
  Helm chart's values.yaml default of loadBalancerOpts.defaultScheme).
  Earlier doc text incorrectly said 'Internal' (that was the dev-machine
  /etc/ override, not the chart default).

- golangci-lint pass: 0 issues across the repo.
  Fixes:
  * pkg/gateway/synth_pool.go      ignore fmt.Fprintf err to *Hash
  * pkg/gateway/gatewayapi_utils.go extract policyCmpEqual / Starts /
                                    Regex constants (replaces 3 inline
                                    'EQUAL_TO' literals flagged by goconst)
  * internal/usecase/gateway_uc/alb_gateway_uc/route_attach.go
                                    pre-allocate matchedParentRefs slice
  * internal/repository/vngcloud_repo/vngcloud_mocks/vngcloud_mock.go
                                    factor real-API error message into a
                                    const + nolint:staticcheck (the trailing
                                    period mirrors upstream wording verbatim)
  * test files: nolint:unparam directives for helpers kept generic on
                                    purpose; _ = on Close/Remove return values
  * gatewayclass_controller.go     drop redundant string(ref.Name) cast
  * route_resolve_test.go          drop redundant int32(port) cast

go vet ./...                    clean
go test ./...                   ok (incl. envtest 12s)
golangci-lint run ./...         0 issues

+++ internal/usecase/gateway_uc/alb_gateway_uc/route_resolve_test.go
@@ -19,12 +19,13 @@ import (
 	"github.com/vngcloud/vngcloud-load-balancer-controller/pkg/utils"
 )
 
+//nolint:unparam // test helper kept generic for future cross-group/-kind cases
 func backendRef(group, kind, ns, name string, port int32, weight *int32) gwv1.HTTPBackendRef {
 	br := gwv1.HTTPBackendRef{
 		BackendRef: gwv1.BackendRef{
 			BackendObjectReference: gwv1.BackendObjectReference{
 				Name: gwv1.ObjectName(name),
-				Port: (*gwv1.PortNumber)(ptr.To(int32(port))),
+				Port: (*gwv1.PortNumber)(ptr.To(port)),
 			},
 			Weight: weight,
 		},
